+++ src/Raven.CodeAnalysis/Workspaces/ProjectSystem/MsBuildProjectSystemService.cs
@@ -128,8 +128,10 @@ public sealed class MsBuildProjectSystemService : IProjectSystemService
                 evaluation.PackageReferences,
                 evaluation.FrameworkReferences);
 
-            foreach (var packageReference in packageReferences)
+            foreach (var packageReference in packageReferences.MetadataReferences)
                 solution = solution.AddMetadataReference(projectId, packageReference);
+            foreach (var macroReference in packageReferences.MacroReferences)
+                solution = solution.AddMacroReference(projectId, macroReference);
         }
 
         foreach (var referencedProjectPath in evaluation.ProjectReferencePaths)